requirements.txt
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0